requirements.txt
selenium
webdriver-manager
pillow
pytesseract